Add initial CIQ source files
This change updates the NordicThingy52 sample app from the CIQ SDK to
use the UUIDs for the custom service and characteristic we are using
for the nRF52 DK. Verified this builds and the simulator starts for the
Forerunner 265s. This has not yet been verified on hardware with the
nRF52 DK, though.