Add LED data characteristic support
Add support for a new characteristic DUKE_LED_CHARACTERISTIC in the
DUKE_CUSTOM_SERVICE. The characteristic contains data for the dev-kit
LEDS, currently only LED_4. The characteristic is implemented as read
and write. This initial support does not add any UI elements as that
will be added in a subsequent patch.
Clean up a few debug prints.