Add LED UI button and write BLE button data
Add a UI button that represents LED_4 on the dev-kit. The button is
implemented with a crude check of the onTap function's coordinates.
A better implementation would be to use a Selectable Button (see
https://developer.garmin.com/connect-iq/core-topics/input-handling/).

The button updates it's display state when BUTTON_4 is pressed on the
dev-kit to match the updating LED_4 state. When the on-screen button
is tapped, the updated state is sent via BLE to the dev-kit to then
update LED_4's state.