Add support for GPIOs 11 (output) and 12 (input)
Add support for two GPIO pins in the custom GPIO BLE characteristic.
Pin 11 is configured as an output, it has a UI button that controls
its state.
Pin 12 is configured as an input with an active pullup, it has text to
indicate its state.
Updated layout on main DeviceView to fit new UI fields and better flow.